Demo Api Request works

+++ lib/model/api/main_data.dart
@@ -19,11 +19,12 @@ class MainData {
 
   factory MainData.fromJson(Map<String, dynamic> json) {
     return MainData(
-        temp: json['temp'],
-        feelsLike: json['feels_like'],
-        tempMin: json['temp_min'],
-        tempMax: json['temp_max'],
-        pressure: json['pressure'],
-        humidity: json['humidity']);
+        temp: json['temp'].toDouble(),
+        feelsLike: json['feels_like'].toDouble(),
+
+        tempMin: json['temp_min'].toDouble(),
+        tempMax: json['temp_max'].toDouble(),
+        pressure: json['pressure'].toDouble(),
+        humidity: json['humidity'].toDouble());
   }
 }
