+++ lib/model/api/wind.dart
@@ -10,6 +10,6 @@ class Wind {
   });
 
   factory Wind.fromJson(Map<String, dynamic> json) {
-    return Wind(speed: json['speed'], degree: json['deg']);
+    return Wind(speed: json['speed'].toDouble(), degree: json['deg'].toDouble());
   }
 }
